requirements.txt
flask==2.0.1
numpy==1.21.0
Pillow==8.3.1
scikit-learn==0.24.2 